+++ src/renderer/component/splash/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { selectDaemonVersionMatched } from 'redux/selectors/app';
 import { selectNotification } from 'lbry-redux';
-import { doCheckDaemonVersion } from 'redux/actions/app';
+import { doCheckDaemonVersion, doNotifyUnlockWallet } from 'redux/actions/app';
 import SplashScreen from './view';
 
 const select = state => ({
@@ -11,6 +11,10 @@ const select = state => ({
 
 const perform = dispatch => ({
   checkDaemonVersion: () => dispatch(doCheckDaemonVersion()),
+  notifyUnlockWallet: () => dispatch(doNotifyUnlockWallet()),
 });
 
-export default connect(select, perform)(SplashScreen);
+export default connect(
+  select,
+  perform
+)(SplashScreen);